Implemented showStats method.
Method feches all the questions, answers and userAnswers for a given
inquiry from the database.

After the method finishes the bean will have an ArrayList with the
inqueries Questions.
An ArrayList of ArrayLists containing the answers for each question.
An ArrayList of ArrayLists containing the user answers for each answer.